[wrangler] Move d1/ commands to src/commands/d1/ (Phase 2b)
Move D1 database commands into the src/commands/ directory structure.

Files moved (26 files):
- src/d1/*.ts → src/commands/d1/*.ts
- src/d1/migrations/*.ts → src/commands/d1/migrations/*.ts
- src/d1/timeTravel/*.ts → src/commands/d1/timeTravel/*.ts

Imports updated in:
- src/index.ts
- src/cli.ts
- src/deployment-bundle/bindings.ts
- src/__tests__/d1/*.ts
- src/__tests__/provision.test.ts

No logic changes - only file moves and import updates.
All 2827 tests pass.
