+++ packages/wrangler/src/__tests__/d1/timeTravel.test.ts
@@ -2,7 +2,7 @@ import { COMPLIANCE_REGION_CONFIG_UNKNOWN } from "@cloudflare/workers-utils";
 import { writeWranglerConfig } from "@cloudflare/workers-utils/test-helpers";
 import { http, HttpResponse } from "msw";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { throwIfDatabaseIsAlpha } from "../../d1/timeTravel/utils";
+import { throwIfDatabaseIsAlpha } from "../../commands/d1/timeTravel/utils";
 import { mockAccountId, mockApiToken } from "../helpers/mock-account-id";
 import { mockConsoleMethods } from "../helpers/mock-console";
 import { useMockIsTTY } from "../helpers/mock-istty";
