+++ packages/wrangler/src/__tests__/d1/trimmer.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { mayContainTransaction, trimSqlQuery } from "../../d1/trimmer";
+import { mayContainTransaction, trimSqlQuery } from "../../commands/d1/trimmer";
 
 describe("mayContainTransaction()", () => {
 	it("should return false if there for regular queries", () => {
