+++ packages/wrangler/src/__tests__/provision.test.ts
@@ -26,7 +26,7 @@ import { mswListNewDeploymentsLatestFull } from "./helpers/msw/handlers/versions
 import { runInTempDir } from "./helpers/run-in-tmp";
 import { runWrangler } from "./helpers/run-wrangler";
 import { writeWorkerSource } from "./helpers/write-worker-source";
-import type { DatabaseInfo } from "../d1/types";
+import type { DatabaseInfo } from "../commands/d1/types";
 
 vi.mock("../utils/fetch-secrets", () => ({
 	fetchSecrets: async () => [],
