+++ packages/wrangler/src/commands/d1/delete.ts
@@ -1,10 +1,10 @@
 import chalk from "chalk";
-import { fetchResult } from "../cfetch";
-import { createCommand } from "../core/create-command";
-import { confirm } from "../dialogs";
-import { logger } from "../logger";
-import { requireAuth } from "../user";
-import { printResourceLocation } from "../utils/is-local";
+import { fetchResult } from "../../cfetch";
+import { createCommand } from "../../core/create-command";
+import { confirm } from "../../dialogs";
+import { logger } from "../../logger";
+import { requireAuth } from "../../user";
+import { printResourceLocation } from "../../utils/is-local";
 import { getDatabaseByNameOrBinding } from "./utils";
 import type { Database } from "./types";
