+++ packages/wrangler/src/commands/d1/execute.ts
@@ -14,13 +14,13 @@ import chalk from "chalk";
 import md5File from "md5-file";
 import { Miniflare } from "miniflare";
 import { fetch } from "undici";
-import { fetchResult } from "../cfetch";
-import { createCommand } from "../core/create-command";
-import { getLocalPersistencePath } from "../dev/get-local-persistence-path";
-import { confirm } from "../dialogs";
-import { logger } from "../logger";
-import { readableRelative } from "../paths";
-import { requireAuth } from "../user";
+import { fetchResult } from "../../cfetch";
+import { createCommand } from "../../core/create-command";
+import { getLocalPersistencePath } from "../../dev/get-local-persistence-path";
+import { confirm } from "../../dialogs";
+import { logger } from "../../logger";
+import { readableRelative } from "../../paths";
+import { requireAuth } from "../../user";
 import splitSqlQuery from "./splitter";
 import { getDatabaseByNameOrBinding, getDatabaseInfoFromConfig } from "./utils";
 import type {
