+++ packages/wrangler/src/commands/d1/export.ts
@@ -5,12 +5,12 @@ import { APIError, configFileName, UserError } from "@cloudflare/workers-utils";
 import chalk from "chalk";
 import { Miniflare } from "miniflare";
 import { fetch } from "undici";
-import { fetchResult } from "../cfetch";
-import { createCommand } from "../core/create-command";
-import { getLocalPersistencePath } from "../dev/get-local-persistence-path";
-import { logger } from "../logger";
-import { readableRelative } from "../paths";
-import { requireAuth } from "../user";
+import { fetchResult } from "../../cfetch";
+import { createCommand } from "../../core/create-command";
+import { getLocalPersistencePath } from "../../dev/get-local-persistence-path";
+import { logger } from "../../logger";
+import { readableRelative } from "../../paths";
+import { requireAuth } from "../../user";
 import { getDatabaseByNameOrBinding, getDatabaseInfoFromConfig } from "./utils";
 import type { Database, ExportPollingResponse, PollingFailure } from "./types";
 import type { ComplianceConfig, Config } from "@cloudflare/workers-utils";
