+++ packages/wrangler/src/commands/d1/index.ts
@@ -1,4 +1,4 @@
-import { createNamespace } from "../core/create-command";
+import { createNamespace } from "../../core/create-command";
 
 export const d1Namespace = createNamespace({
 	metadata: {
