+++ packages/wrangler/src/commands/d1/insights.ts
@@ -1,7 +1,7 @@
-import { fetchGraphqlResult } from "../cfetch";
-import { createCommand } from "../core/create-command";
-import { logger } from "../logger";
-import { requireAuth } from "../user";
+import { fetchGraphqlResult } from "../../cfetch";
+import { createCommand } from "../../core/create-command";
+import { logger } from "../../logger";
+import { requireAuth } from "../../user";
 import {
 	getDatabaseByNameOrBinding,
 	getDatabaseInfoFromIdOrName,
