+++ packages/wrangler/src/commands/d1/list.ts
@@ -1,7 +1,7 @@
-import { fetchResult } from "../cfetch";
-import { createCommand } from "../core/create-command";
-import { logger } from "../logger";
-import { requireAuth } from "../user";
+import { fetchResult } from "../../cfetch";
+import { createCommand } from "../../core/create-command";
+import { logger } from "../../logger";
+import { requireAuth } from "../../user";
 import type { Database } from "./types";
 import type { ComplianceConfig } from "@cloudflare/workers-utils";
