+++ packages/wrangler/src/commands/d1/trimmer.ts
